Navigation class moved to modules
