+++ js/index.js
@@ -1,23 +1,4 @@
-import Navigation from "./Navigation.js";
-
-const nav = document.getElementById('nav');
 const footerYear = document.getElementById('footer-year');
-
 const date = new Date().getFullYear();
 //Insert the year
 footerYear.innerText = date.toString();
-
-//Creates an array with the ids
-const fillElementsList = () => {
-    let navElements = [...nav.children];
-    return navElements.map(element => element.id);
-}
-
-const navigation = new Navigation(fillElementsList());
-
-//Handle the click events in nav
-const handleNavClick = evt => {
-    navigation.renderMessage(evt.target.id);
-}
-
-nav.addEventListener('click', handleNavClick);
